// src/app/modules/shared/models/companysignup.ts
// Create a file named company-form.model.ts

export interface CompanyForm {
    companyName: string;
    owner: string;
    streetAddress: string;
    city: string;
    state: string;
    zipcode: string;
    country: string;
    email: string;
    phoneNumber: string;
    nominalCapital: number;
    industry: string;
    website: string;
    companyDescription: string;
    verificationCheckbox: boolean;
  }
  